#4663 refactoring * merge 2 ad_element dao classes together
https://github.com/metasfresh/metasfresh/issues/4664
