+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/ad/element/api/impl/ADElementDAO.java
@@ -1,7 +1,4 @@
-/**
- *
- */
-package org.adempiere.ad.service.impl;
+package org.adempiere.ad.element.api.impl;
 
 import static org.adempiere.model.InterfaceWrapperHelper.loadOutOfTrx;
 import static org.adempiere.model.InterfaceWrapperHelper.newInstance;
@@ -15,7 +12,7 @@ import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.ad.dao.impl.UpperCaseQueryFilterModifier;
 import org.adempiere.ad.element.api.AdElementId;
 import org.adempiere.ad.element.api.CreateADElementRequest;
-import org.adempiere.ad.service.IADElementDAO;
+import org.adempiere.ad.element.api.IADElementDAO;
 import org.adempiere.ad.table.api.IADTableDAO;
 import org.adempiere.util.LegacyAdapters;
 import org.compiere.model.I_AD_Column;
@@ -23,19 +20,50 @@ import org.compiere.model.I_AD_Element;
 import org.compiere.model.I_AD_Field;
 import org.compiere.model.I_AD_Menu;
 import org.compiere.model.I_AD_Tab;
+import org.compiere.model.I_AD_UI_Element;
 import org.compiere.model.I_AD_Window;
 import org.compiere.model.MColumn;
 
 import de.metas.util.Services;
 import lombok.NonNull;
 
-/**
- * @author cg
+/*
+ * #%L
+ * de.metas.adempiere.adempiere.base
+ * %%
+ * Copyright (C) 2017 metas GmbH
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 2 of the
+ * License, or (at your option) any later version.
  *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public
+ * License along with this program. If not, see
+ * <http://www.gnu.org/licenses/gpl-2.0.html>.
+ * #L%
  */
-// TODO: merge into ElementDAO
+
 public class ADElementDAO implements IADElementDAO
 {
+
+	@Override
+	public List<I_AD_UI_Element> retrieveChildUIElements(final I_AD_Element element)
+	{
+		return Services.get(IQueryBL.class).createQueryBuilder(I_AD_Element.class)
+				.addEqualsFilter(I_AD_Element.COLUMN_AD_Element_ID, element.getAD_Element_ID())
+				.andCollectChildren(I_AD_Column.COLUMN_AD_Element_ID)
+				.andCollectChildren(I_AD_Field.COLUMN_AD_Column_ID)
+				.andCollectChildren(I_AD_UI_Element.COLUMN_AD_Field_ID)
+				.create()
+				.list();
+
+	}
 
 	@Override
 	public List<I_AD_Column> retrieveColumns(final int elementId)
@@ -139,4 +167,7 @@ public class ADElementDAO implements IADElementDAO
 
 		return AdElementId.ofRepoId(record.getAD_Element_ID());
 	}
+
+
+
 }
