+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/ad/element/model/interceptor/AD_Element.java
@@ -2,9 +2,9 @@ package org.adempiere.ad.element.model.interceptor;
 
 import java.util.List;
 
+import org.adempiere.ad.element.api.IADElementDAO;
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
-import org.adempiere.ad.service.IADElementDAO;
 import org.adempiere.exceptions.AdempiereException;
 import org.compiere.model.I_AD_Column;
 import org.compiere.model.I_AD_Element;
@@ -23,12 +23,12 @@ import de.metas.util.Services;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -41,6 +41,8 @@ public class AD_Element
 	@ModelChange(timings = { ModelValidator.TYPE_BEFORE_CHANGE }, ifColumnsChanged = I_AD_Element.COLUMNNAME_ColumnName)
 	public void onColumnNameDeleted(final I_AD_Element element)
 	{
+		final IADElementDAO adElementDAO = Services.get(IADElementDAO.class);
+
 		final String columnName = element.getColumnName();
 		if (!Check.isEmpty(columnName))
 		{
@@ -48,7 +50,8 @@ public class AD_Element
 			return;
 		}
 
-		final List<I_AD_Column> columnsUsingElement = Services.get(IADElementDAO.class).retrieveColumns(element.getAD_Element_ID());
+
+		final List<I_AD_Column> columnsUsingElement = adElementDAO.retrieveColumns(element.getAD_Element_ID());
 
 		if (!Check.isEmpty(columnsUsingElement))
 		{
