+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/ad/menu/model/interceptor/AD_Menu.java
@@ -6,10 +6,10 @@ import org.adempiere.ad.callout.annotations.Callout;
 import org.adempiere.ad.callout.annotations.CalloutMethod;
 import org.adempiere.ad.callout.spi.IProgramaticCalloutProvider;
 import org.adempiere.ad.element.api.AdElementId;
+import org.adempiere.ad.element.api.IADElementDAO;
 import org.adempiere.ad.modelvalidator.annotations.Init;
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
-import org.adempiere.ad.service.IADElementDAO;
 import org.compiere.model.I_AD_Element;
 import org.compiere.model.I_AD_Menu;
 import org.compiere.model.I_AD_Tab;
@@ -56,6 +56,7 @@ public class AD_Menu
 	@CalloutMethod(columnNames = I_AD_Menu.COLUMNNAME_AD_Element_ID)
 	public void onElementIDChanged(final I_AD_Menu menu) throws SQLException
 	{
+		final IADElementDAO adElementDAO = Services.get(IADElementDAO.class);
 
 		if (!IElementTranslationBL.DYNATTR_AD_Menu_UpdateTranslations.getValue(menu, true))
 		{
@@ -63,7 +64,7 @@ public class AD_Menu
 			return;
 		}
 
-		final I_AD_Element menuElement = Services.get(IADElementDAO.class).getById(menu.getAD_Element_ID());
+		final I_AD_Element menuElement = adElementDAO.getById(menu.getAD_Element_ID());
 
 		if (menuElement == null)
 		{
@@ -81,7 +82,6 @@ public class AD_Menu
 	@ModelChange(timings = { ModelValidator.TYPE_AFTER_NEW, ModelValidator.TYPE_AFTER_CHANGE }, ifColumnsChanged = I_AD_Menu.COLUMNNAME_AD_Element_ID)
 	public void updateTranslationsForElement(final I_AD_Menu menu)
 	{
-
 		final AdElementId menuElementId = AdElementId.ofRepoIdOrNull(menu.getAD_Element_ID());
 		if (menuElementId == null)
 		{
